backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
google-generativeai==0.8.3
google-cloud-speech==2.27.0
google-cloud-texttospeech==2.17.2
playwright==1.48.0
python-slugify==8.0.4
httpx==0.27.2
